feat: Add WithAuthOnly component for  access control
This commit adds a new component called WithAuthOnly, which is used for controlling access to certain parts of the application based on whether the user is signed in or not. It takes in two props: children and alternative. If the user is not signed in, the alternative content is rendered instead of the children. This component is used in the Auth Debug page to display a card with different content depending on the user's authentication status.