add:CopyToClipboard component
This commit adds a new component called CopyToClipboard. It is a button component that allows users to copy content to their clipboard. The component uses the Clipboard API to write the content to the clipboard when clicked. It also provides visual feedback to the user by showing a success icon when the content is successfully copied.